The sequence counter class is placed in a separate translation unit + fixed second maximum function
