+++ nikitov.dmitriy/P1/main.cpp
@@ -1,35 +1,15 @@
 #include <iostream>
-#include <limits>
+#include "sequence_counter.cpp"
 
 void isMax(int now, int& actual, int& pre)
 {
   if (now >= actual){
     pre = actual;
     actual = now;
+  } else if (now > pre){
+    pre = now;
   }
 }
-
-struct SequenceCounter
-{
-  SequenceCounter():
-    count_(0)
-   {}
-
-  void isLimit(int)
-  {
-    size_t max_limit = std::numeric_limits<size_t>::max();
-    if (count_ == max_limit){
-      throw std::logic_error("This sequence has no end\n");
-    }
-    ++count_;
-  }
-  size_t check_count()
-  {
-    return count_;
-  }
-  private:
-    size_t count_;
-};
 
 int main()
 {
@@ -39,6 +19,7 @@ int main()
   int pre_max = 0;
   int &pre = pre_max;
   SequenceCounter counter;
+
   do{
     std::cin >> number;
     if (!std::cin){
@@ -56,8 +37,9 @@ int main()
     }
   }
   while (number != 0);
+
   if (counter.check_count() == 0 || counter.check_count() == 1){
-    std::cerr << "Error This sequence is too short\n";
+    std::cerr << "Error: This sequence is too short\n";
     return 2;
   }
   std::cout << pre_max << '\n';
